Deliverables/Code/event-schedular-saavyshopper/requirements.txt
APScheduler==3.10.1
async-timeout==4.0.2
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
firebase-admin==6.2.0
Flask==2.3.2
Flask-Cors==4.0.0
flask-localtunnel==1.0.6
flask-marshmallow==0.15.0
google-api-core==2.11.1
google-api-python-client==2.92.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.19.0
msgpack==1.0.5
packaging==23.1
proto-plus==1.22.3
protobuf==4.23.4
py-localtunnel==1.0.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.1.0
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
rsa==4.9
six==1.16.0
soupsieve==2.4.1
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.6
zipp==3.15.0
